pyproject.toml
[tool.poetry]
name = "strawberry-graphql-permission-extension"
version = "0.1.0"
description = ""
authors = ["Alteian <Alteian@proton.me>"]
readme = "README.md"
packages = [{include = "strawberry_graphql_permission_extension"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^3.2"
strawberry-graphql = "^0.195.2"
strawberry-graphql-django = "^0.12.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
